Refactor iXBRLViewer functionality into a class.
